fix code review
