+++ src/theme.js
@@ -2,9 +2,9 @@ import { createMuiTheme }  from '@material-ui/core/styles'
 
 const theme = createMuiTheme({
   palette: {
-    primary: { main: '#00C6CF',contrastText:'#fff' },
-    secondary: { main: '#ff5f5f' },
-  },
+    primary: { main: '#00C6CF', contrastText: '#fff' },
+    secondary: { main: '#ff5f5f' }
+  }
 })
 
 export default theme
