+++ src/views/History.jsx
@@ -49,27 +49,16 @@ class History extends Component {
           }}
           localization={{ body: { editRow: { deleteText: 'Are you sure you want to delete this log?' } } }}
           editable={{
-            onRowUpdate: (newData, oldData) =>
-              new Promise((resolve, reject) => {
-                setTimeout(() => {
-                  console.log("update")
-                  resolve();
-                }, 1000);
-              })
-            ,
             onRowDelete: oldData =>
             new Promise((resolve, reject) => {
-              setTimeout(() => {
-                this.props.removeLog(
-                  localStorage.getItem("uid"),
-                  null,
-                  oldData.id
-                )
-                resolve();
-              }, 1000);
+              this.props.removeLog(
+                localStorage.getItem("uid"),
+                null,
+                oldData.id
+              )
+              resolve();
             })
-        }}
-
+          }}
         />
       </div>
     );
