+++ src/views/UserProfile.css
@@ -7,12 +7,6 @@
   min-height: 230px;
 }
 
-.profile-form {
-  margin: 10px;
-  width: 10vw;
-  height: 10vw;
-}
-
 p {
   color: grey;
   margin: 3px;
